requirements.txt
# Core frameworks - using latest compatible versions
langchain==0.3.13
langgraph==0.2.50
langchain-community==0.3.13
langchain-groq==0.2.1
langchain-core>=0.3.26

# Fix for Python 3.13 compatibility
langsmith>=0.1.147

# Vector store & embeddings - using new packages
chromadb==0.5.23
langchain-chroma==0.1.4
langchain-huggingface==0.1.2
sentence-transformers==3.3.1

# Document processing
pypdf==5.1.0
python-docx==1.1.2
docx2txt==0.8

# APIs
groq==0.13.0
google-search-results==2.4.2

# UI
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
tenacity==9.0.0

# Logging & monitoring
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-cov==6.0.0
